+++ crates/nu-utils/src/lib.rs
@@ -1,6 +1,7 @@
 mod casing;
 pub mod ctrl_c;
 mod deansi;
+pub mod emoji;
 pub mod locale;
 pub mod utils;
 
@@ -14,3 +15,4 @@ pub use casing::IgnoreCaseExt;
 pub use deansi::{
     strip_ansi_likely, strip_ansi_string_likely, strip_ansi_string_unlikely, strip_ansi_unlikely,
 };
+pub use emoji::is_emoji;
